Removed template parameter from STK field containers.
It was never used to deduce compile-time information.
The 3 instantiations of the same class are now one.
