+++ src/disc/stk/Albany_MultiSTKFieldContainer.hpp
@@ -12,27 +12,28 @@
 
 namespace Albany {
 
-template <DiscType Interleaved>
-class MultiSTKFieldContainer : public GenericSTKFieldContainer<Interleaved>
+class MultiSTKFieldContainer : public GenericSTKFieldContainer
 {
  public:
   MultiSTKFieldContainer(
-      const Teuchos::RCP<Teuchos::ParameterList>&               params_,
-      const Teuchos::RCP<stk::mesh::MetaData>&                  metaData_,
-      const Teuchos::RCP<stk::mesh::BulkData>&                  bulkData_,
-      const int                                                 numDim_,
-      const Teuchos::RCP<StateInfoStruct>&                      sis,
-      const int                                                 num_params);
+      const Teuchos::RCP<Teuchos::ParameterList>&     params_,
+      const Teuchos::RCP<stk::mesh::MetaData>&        metaData_,
+      const Teuchos::RCP<stk::mesh::BulkData>&        bulkData_,
+      const DiscType                                  interleaved_,
+      const int                                       numDim_,
+      const Teuchos::RCP<StateInfoStruct>&            sis,
+      const int                                       num_params);
  
   MultiSTKFieldContainer(
-      const Teuchos::RCP<Teuchos::ParameterList>&               params_,
-      const Teuchos::RCP<stk::mesh::MetaData>&                  metaData_,
-      const Teuchos::RCP<stk::mesh::BulkData>&                  bulkData_,
-      const int                                                 neq_,
-      const int                                                 numDim_,
-      const Teuchos::RCP<Albany::StateInfoStruct>&              sis,
-      const Teuchos::Array<Teuchos::Array<std::string>>&        solution_vector,
-      const int                                                 num_params);
+      const Teuchos::RCP<Teuchos::ParameterList>&           params_,
+      const Teuchos::RCP<stk::mesh::MetaData>&              metaData_,
+      const Teuchos::RCP<stk::mesh::BulkData>&              bulkData_,
+      const DiscType                                        interleaved_,
+      const int                                             neq_,
+      const int                                             numDim_,
+      const Teuchos::RCP<Albany::StateInfoStruct>&          sis,
+      const Teuchos::Array<Teuchos::Array<std::string>>&    solution_vector,
+      const int                                             num_params);
 
   ~MultiSTKFieldContainer() = default;
