+++ src/disc/stk/Albany_OrdinarySTKFieldContainer.hpp
@@ -11,8 +11,7 @@
 
 namespace Albany {
 
-template <DiscType Interleaved>
-class OrdinarySTKFieldContainer : public GenericSTKFieldContainer<Interleaved>
+class OrdinarySTKFieldContainer : public GenericSTKFieldContainer
 {
  public:
   OrdinarySTKFieldContainer(
@@ -20,6 +19,7 @@ class OrdinarySTKFieldContainer : public GenericSTKFieldContainer<Interleaved>
       const Teuchos::RCP<stk::mesh::MetaData>&                  metaData_,
       const Teuchos::RCP<stk::mesh::BulkData>&                  bulkData_,
       const AbstractFieldContainer::FieldContainerRequirements& req,
+      const DiscType                                            interleaved_,
       const int                                                 numDim_,
       const Teuchos::RCP<Albany::StateInfoStruct>&              sis,
       const int                                                 num_params);
@@ -28,6 +28,7 @@ class OrdinarySTKFieldContainer : public GenericSTKFieldContainer<Interleaved>
       const Teuchos::RCP<Teuchos::ParameterList>&               params_,
       const Teuchos::RCP<stk::mesh::MetaData>&                  metaData_,
       const Teuchos::RCP<stk::mesh::BulkData>&                  bulkData_,
+      const DiscType                                            interleaved_,
       const int                                                 neq_,
       const AbstractFieldContainer::FieldContainerRequirements& req,
       const int                                                 numDim_,
